Fix path error calling functions from main.py
We want to call functions using the cli tool. All necessary tools needed for the pipeline should be stored in the src folder and the paths should be adjusted accordingly. We added flexible ROOT variables to mothur_process.py to create the correct paths being either called from main.py or when mothur_process is run stand alone.